move dep under xpkg
