+++ internal/xpkg/dep/cache/cache_test.go
@@ -24,7 +24,7 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
-	"github.com/upbound/up/internal/dep"
+	"github.com/upbound/up/internal/xpkg/dep"
 
 	ociname "github.com/google/go-containerregistry/pkg/name"
 	v1 "github.com/google/go-containerregistry/pkg/v1"
