+++ internal/xpkg/dep/resolver.go
