+++ internal/xpkg/dep/resolver_test.go
