+++ internal/xpkg/dep/workspace_test.go
